Register and clear up hook_manager within session lifecycle
